Initial files, Bubble sort and count sort
Bubble sort - Can be used for any mutables. This case uses files for
input and output. Can easily be pulled out for use in another form.
Incredibly slow for large amounts of data.

Count sort - Can be used for any integers. Again this uses file input
and output and can be easily used again. Might not be the most
optimized, but it works much faster than the bubble sort.